started creating a rough firmaware
Numpad and macro layout made, rgb and oled display started with Hermes logo on oled display. It is needed to configure an interface stating layer, caps, rgb state and maybe more info on the oled display. RGB needs modes and layers, maybe a way to control the colors with the rotary encoders.